(#300) Refactored most Plates Extensions
