+++ src/Modules/Plates/Extensions/Helpers.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Tapestry\Plates\Extensions;
+namespace Tapestry\Modules\Plates\Extensions;
 
 use League\Plates\Engine;
+use Tapestry\Entities\ProjectFile;
 use Tapestry\Entities\ViewFileTrait;
 use League\Plates\Extension\ExtensionInterface;
